+++ Unigram/Unigram/ViewModels/Chats/ChatSharedMediaViewModel.cs
@@ -528,7 +528,7 @@ namespace Unigram.ViewModels.Chats
         {
         }
 
-        public void OpenUser(int userId)
+        public void OpenUser(long userId)
         {
         }
 
@@ -540,7 +540,7 @@ namespace Unigram.ViewModels.Chats
         {
         }
 
-        public void OpenViaBot(int viaBotUserId)
+        public void OpenViaBot(long viaBotUserId)
         {
         }
