+++ Unigram/Unigram/ViewModels/DialogViewModel.Delegate.cs
@@ -424,7 +424,7 @@ namespace Unigram.ViewModels
             }
         }
 
-        public async void OpenUser(int userId)
+        public async void OpenUser(long userId)
         {
             var response = await ProtoService.SendAsync(new CreatePrivateChat(userId, false));
             if (response is Chat chat)
@@ -441,7 +441,7 @@ namespace Unigram.ViewModels
             }
         }
 
-        public void OpenViaBot(int viaBotUserId)
+        public void OpenViaBot(long viaBotUserId)
         {
             var chat = Chat;
             if (chat != null && chat.Type is ChatTypeSupergroup super && super.IsChannel)
@@ -647,7 +647,7 @@ namespace Unigram.ViewModels
             return null;
         }
 
-        public string GetAdminTitle(int userId)
+        public string GetAdminTitle(long userId)
         {
             var chat = _chat;
             if (chat == null)
