+++ Unigram/Unigram/ViewModels/Drawers/AnimationDrawerViewModel.cs
@@ -275,7 +275,7 @@ namespace Unigram.ViewModels.Drawers
         private readonly IProtoService _protoService;
         private readonly string _query;
 
-        private int? _userId;
+        private long? _userId;
         private string _offset = string.Empty;
         private bool _hasMoreItems = true;
