+++ Unigram/Unigram/Views/Popups/SupergroupEventLogFiltersPopup.xaml.cs
@@ -16,7 +16,7 @@ namespace Unigram.Views.Popups
         private ICacheService _cacheService;
 
         public ChatEventLogFilters Filters { get; private set; }
-        public IList<int> UserIds { get; private set; }
+        public IList<long> UserIds { get; private set; }
 
         public SupergroupEventLogFiltersPopup()
         {
@@ -27,7 +27,7 @@ namespace Unigram.Views.Popups
             SecondaryButtonText = Strings.Resources.Cancel;
         }
 
-        public Task<ContentDialogResult> ShowAsync(IProtoService protoService, int supergroupId, ChatEventLogFilters filters, IList<int> userIds)
+        public Task<ContentDialogResult> ShowAsync(IProtoService protoService, long supergroupId, ChatEventLogFilters filters, IList<long> userIds)
         {
             _protoService = protoService;
             _cacheService = protoService;
@@ -166,7 +166,7 @@ namespace Unigram.Views.Popups
             };
 
             var areAllAdministratorsSelected = List.Items.All(x => List.SelectedItems.Contains(x));
-            UserIds = areAllAdministratorsSelected ? new int[0] : List.SelectedItems.OfType<ChatMember>().Select(x => x.MemberId).OfType<MessageSenderUser>().Select(x => x.UserId).ToArray();
+            UserIds = areAllAdministratorsSelected ? new long[0] : List.SelectedItems.OfType<ChatMember>().Select(x => x.MemberId).OfType<MessageSenderUser>().Select(x => x.UserId).ToArray();
         }
 
         private void ContentDialog_SecondaryButtonClick(ContentDialog sender, ContentDialogButtonClickEventArgs args)
